Create ChannelOpen actor
We create a new actor that will handle the channel opening flow.
In this commit we simply copy-paste all of the corresponding channel code.

This commit doesn't even build, but it will make it easier to review the
actual logic changes in the following commits.
